Commit_12 AdminController and save methods added.

+++ src/main/resources/db/migration/V1_4__init_rent_a_car_app_car_to_rent.sql
@@ -6,7 +6,7 @@ CREATE TABLE car_to_rent
     car_type        VARCHAR(30)    NOT NULL,
     brand           VARCHAR(30)    NOT NULL,
     model           VARCHAR(30)    NOT NULL,
-    production_year SMALLINT       NOT NULL,
+    production_year INT       NOT NULL,
     color           VARCHAR(30)    NOT NULL,
     car_status      VARCHAR(30)    NOT NULL,
